Add list of skills to homepage
A new section was added to the homepage that displays a list of
the technologies I'm familiar with. Specifically, it displays
the official logo of each technology (i.e., the React logo, the
Vue.js logo, etc.). As a result, several SVG files were added
to the /assets/images/logos directory.